% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{every_nth}
\alias{every_nth}
\title{Function for minor tick marks}
\usage{
every_nth(x, nth, empty = TRUE, inverse = FALSE)
}
\arguments{
\item{x}{.}

\item{nth}{.}

\item{empty}{.}

\item{inverse}{.}
}
\value{
.
}

