% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{insert_minor}
\alias{insert_minor}
\title{Insert minor tick marks}
\usage{
insert_minor(major_labs, n_minor, multiple)
}
\arguments{
\item{major_labs}{.}

\item{n_minor}{.}

\item{multiple}{.}
}
\value{
.
}

