% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Plot data}
\alias{Plot data}
\alias{smps_plot}
\title{Plot the data}
\usage{
smps_plot(the_prepared_data, aspect_ratio = 1.1/5, font_size = 12,
  h = "2 hours", y_breaks = 100, colour_ramp = colour_ramp_igor,
  y_axis = "linear", seconds_offset = 1000, ...)
}
\arguments{
\item{the_prepared_data}{the output from the prepare_data function}

\item{aspect_ratio}{change the aspect ratio of the plot}

\item{font_size}{change the axis and legend text size}

\item{h}{how many hours apart the x-axis ticks should be}

\item{y_breaks}{location of major ticks on the y axis}

\item{y_axis}{'linear' or 'log' (base 10) y-axis}

\item{seconds_offset}{offset in seconds for adjusting the log y-axis}

\item{...}{so you can send other things to modify the theme and scales}
}

