% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare the data}
\usage{
prepare_data(the_data, cols = 1:ncol(the_data), interval_in_seconds = 200,
  interval_diameter = 0.75)
}
\arguments{
\item{the_data}{data frame with rows as day-time, columns as particle sizes}

\item{cols}{subset of columns to prepare, if the_data has other columns that should be ignored by this function. Default is all columns}

\item{interval_in_seconds}{interval to use when interpolating. Default is 200, higher is faster, but lower resolution on the plot}

\item{interval_diameter}{interval to use when interpolating. Default is 0.75, higher is faster, but lower resolution on the plot}
}

