% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{reformat_date}
\alias{reformat_date}
\title{reformat_date}
\usage{
reformat_date(the_excel_dates)
}
\arguments{
\item{the_dates}{}
}
\description{
Convert Excel format dates to POSIX Date-time values
}
\details{
In Excel if we have a date like this: 16/1/2013 23:58 it's not very
useful in R. This function rearranges the Excel date string into
a POSIX date-time string such as:  "2013-01-16 23:58:00 ICT". This is
useful if you want to add other layers to the main plot and you
need to convert the Excel date format into the POSIX format that
we're using for the main plot.
}

